requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
google-generativeai>=0.5.0
python-multipart==0.0.6

# Supabase (File Storage)
supabase>=2.4.0

# Pinecone (Vector Database)
pinecone-client==3.0.0

# PDF Processing
PyPDF2==3.0.1
pdfplumber==0.10.3

# Image Processing
Pillow==10.1.0

# Audio Processing
SpeechRecognition==3.10.0

# Firebase Admin
firebase-admin==6.2.0

# HTTP & Utilities
requests==2.31.0
aiohttp==3.9.1
numpy==1.26.2
pydantic==2.5.2

# Data Processing
pandas==2.1.3
python-dateutil==2.8.2
